Through Unit Testing, Multiple Files Were Modified In Order To Get Tests To Pass. For The User Profile Class, ReMapped Bill To User Profile & Allow To Save And Load A User Profile. For The User Repository Interface, Added Ability To Find One By User Profile Which Was Needed For The Rest Controller. For The User Profile Rest Controller Test & Class, Created Through Multiple Tests, The Bill Section Was Created & Added.
A lot of re-configuration had to be completed in order for me to make it this far but hour's later, I was finally able to re-write some of the code to accomate changes needed in order to make the tests not only functional but was able to correct failed tests and convert them into passing tests.
Feeling Accomplished. :D